\name{Glib.createDataTransformationManager}
\alias{Glib.createDataTransformationManager}
\title{Glib.createDataTransformationManager}
\usage{
Glib.createDataTransformationManager(env, data)
}
\arguments{
  \item{env}{Glib env object.}
  \item{data}{event log data frame.}
}
\description{
Return an transformation manager object.
}
\examples{
\dontrun{
UserProfilesManager <- Glib.createUserProfilesManager(Glib.environment)
}
}
