\name{Glib.createUserProfilesManager}
\alias{Glib.createUserProfilesManager}
\title{Glib.createUserProfilesManager}
\usage{
Glib.createUserProfilesManager(env, data)
}
\arguments{
  \item{env}{Glib env object.}
  \item{data}{input data frame.}
}
\description{
Return an user profile manager object.
}
\examples{
\dontrun{
UserProfilesManager <- Glib.createUserProfilesManager(Glib.environment)
}
}
