\name{Glib.dropEventLogsAfterGoalEvent}
\alias{Glib.dropEventLogsAfterGoalEvent}
\title{Glib.dropEventLogsAfterGoalEvent}
\usage{
Glib.dropEventLogsAfterGoalEvent(manager)
}
\arguments{
  \item{manager}{Glib data transformation manager object.}
}
\description{
Remove logs after goal event at each user.
}
\examples{
\dontrun{
Glib.dropEventLogsAfterGoalEvent(DataTransformationManager)
}
}
