\name{Glib.getSummaryList}
\alias{Glib.getSummaryList}
\title{Glib.getSummaryList}
\usage{
Glib.getSummaryList(manager)
}
\arguments{
  \item{manager}{Glib user profiles manager object.}
}
\description{
Return the summary list from all events.
}
\examples{
\dontrun{
summaryList <- Glib.getSummaryList(UserProfilesManager)
}
}
