\name{Glib.setUniqueEventOperation}
\alias{Glib.setUniqueEventOperation}
\title{Glib.setUniqueEventOperation}
\usage{
Glib.setUniqueEventOperation(
  manager, event, operation
)
}
\arguments{
  \item{manager}{Glib user profiles manager object.}
  \item{event}{event name.}
  \item{operation}{sum, avg, ...}
}
\description{
...
}
\examples{
\dontrun{
...
}
}
